Update
The about markdown file holds a layout with the forntmatter.
The desktop version, my desktop version, holds also a titel
variable. The titel variable is about not redefined.
It is not possible for me redefine the environment. I asume
the about layout in the gihub file structure with jekyll for git
pages is written in html file format while a markdown about
file is not the layout path in a navigation yml file.
the 404 page output for the solarplexus markdown file content
is just maybe because the navigation link line holds the typed
file extension md and not html.