Update
I have no solution. The desktop version will be enchased with the cayman
theme. If the SEO unknowen output for the zip files would not have been
etcetera. The zip files are a part of definging architecture. I have not studied
files with pif extensions but that is years ago.