Update and rename 2022-08-28-Windows10Umgebungsvariablen.markdown to 2022-08-31-Windows10Umgebungsvariablen.markdown
The aim is to create an zertifikat test suit environment. With the page
https://docs.microsoft.com/en-us/previous-versions/k4cbh4dh(v=vs.140)
i had the oportunity in a VB dot NET web form ASP dot NET to enter the input
 <!--   <div>
    Welcome to Visual Web Developer
    <br />
    </div> -->
    <div>
    Welcome to Visual Web Developer
    <br />
    </div>
the launch first time succeded. The second time a message apeared
	Das Projekt ist für die Verwendung von SSL konfiguriert. Um die Anzeige von
	SSL-Warnungen im Browser zu unterbinden, können Sie auswählen, dass Sie
	dem selbstsignierten, von IIS Express generierten SSL-Zertifikat vertrauen.

	Möchten Sie dem von IIS Express generierten SSL-Zertifikat vertrauen?
i choose no, because secure socket layer may have labels integratet with the system. Since i do not
know how to aim secure socket layer messages with security internet accesses (ports)
to refuse them is the first step i stoped recognizing for me what is possible.
The provided option to get a zertifikat instead is still refused from me. The
message in the browser after the message shown without possibilty to track the source
is Ihre Verbindung ist nicht privat. I do not asume a fraud connection but a maybe provider test.
My try to establish a ergonomic "Ring", to move without money and media connectivity
,
i lead to the extension ZTS zertifikat test suit. Now the comment html part is not obvious not even near filled
with informations related to money typescript to true type or offline asp web form assumptions
with the aim to find jekyll offline connectity points that can be documentet offline for further
aims. The asp dot net web form is MAYBE useable with a BITMAP file for first bridges
with onlie pages. And the jekyll offline pages have to be prepared from me for my possibilties
to use actions in a user agent environment to learn to decide
when, but it is not possible without experinces, livereload has to lead or reaimed
with funds. End.