Create LGE_scale_vector_graphic.svg
It is very possible you can get me wrong.
My native language is not english. True.
scale vector graphics are not main entry points for a network. If i start with a svg extension
that can be exposed to a stream my second step would be to work with the csv extension
common seperator value.
My assumption with a movie name twister first to try recognize what i have measured
and start with a second coma wave.
And as i am not able to aim for myself the next file for a stream has to be prepared.
The main aim is to quit the svg extension file to offline mode. 0 1 End.